safari-534.53-branch/LayoutTests/http/tests/media/text-served-as-text-expected.txt

EXPECTED (video.error == 'null') OK
EVENT(loadstart)
EVENT(error)
failed trying to load text file served as text OK
END OF TEST
